+++ src/components/category/Category.vue
