+++ src/views/CategoryView.vue
@@ -1,5 +1,5 @@
 <script setup>
-import Category from '@/components/Category.vue';
+import Category from '@/components/category/Category.vue';
 </script>
 
 <template>
